+++ ChangeLog.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.2] - 2022-06-29
 ### Added
-- add the `sh` function that does not `IrrepsData` as input/output
+- add the `sh` function that does not use `IrrepsData` as input/output
 - `legendre` algorithm to compute spherical harmonics
 - add flag `algorithm` to specify the algorithm to use for computing spherical harmonics, use `legendre` for large L.
 - `experimental.voxel_convolution`: add optional dynamic steps (not static for jit)